refactor(auth)!: changed route for change password function from /user to /auth
