+++ internal/api/v1/auth/change_password_handler.go
@@ -1,4 +1,4 @@
-package users
+package auth
 
 import (
 	"github.com/gin-gonic/gin"
